refactor: wip!
