+++ examples/rust-mix/ftl-project.toml
@@ -1,0 +1,8 @@
+ftl-min-version = "dev"
+hermit = false
+no-git = false
+
+[global]
+
+[commands]
+  startup = [""]
