+++ examples/rust-mix/gooo/ftl.toml
@@ -1,0 +1,2 @@
+module = "gooo"
+language = "go"
