+++ examples/rust-mix/ruuu/ftl.toml
@@ -1,0 +1,2 @@
+module = "gooo"
+language = "go"
