+++ rust-runtime/ftl/src/lib.rs
@@ -1,7 +1,8 @@
-pub use build::build;
 pub use ftl_derive::verb;
+pub use parser::build;
 
-mod build;
+mod generator;
+mod parser;
 pub mod runner;
 pub mod schema;
 pub mod verb_server;
