+++ rust-runtime/ftl/src/schema.rs
@@ -1,9 +1,10 @@
 //! A crate for parsing/generating code to and from schema binary.
 use proc_macro2::Ident;
 use prost::Message;
-use syn::Path;
 
 use ftl_protos::schema;
+
+use crate::parser;
 
 pub fn binary_to_module(mut reader: impl std::io::Read) -> schema::Module {
     let mut buf = Vec::new();
@@ -13,10 +14,10 @@ pub fn binary_to_module(mut reader: impl std::io::Read) -> schema::Module {
 
 pub fn code_to_module(module: &Ident, code: &str) -> schema::Module {
     let ast = syn::parse_file(code).unwrap();
-    let verbs = extract_ast_verbs(module, ast);
+    let verbs = parser::extract_ast_verbs(module, ast);
 
     let verbs = verbs.iter().filter_map(|verb| {
-        let has_ctx = fn_has_context_as_first_arg(&verb.func);
+        let has_ctx = parser::fn_has_context_as_first_arg(&verb.func);
         if !has_ctx {
             panic!("First argument must be of type Context");
         }
@@ -37,122 +38,6 @@ pub fn code_to_module(module: &Ident, code: &str) -> schema::Module {
         name: "".to_string(),
         decls,
     }
-}
-
-#[derive(Debug)]
-pub struct VerbToken {
-    pub module: Ident,
-    pub func: syn::ItemFn,
-}
-
-impl VerbToken {
-    fn to_proto(&self) -> schema::Verb {
-        let mut verb = schema::Verb::default();
-        verb.name = self.func.sig.ident.to_string();
-
-        let syn::FnArg::Typed(arg) = self.func.sig.inputs.first().unwrap() else {
-            panic!("Function must have at least one argument");
-        };
-
-        let syn::Type::Reference(path) = &*arg.ty else {
-            panic!("First argument must not be a self argument");
-        };
-
-        let syn::Type::Path(type_path) = &*path.elem else {
-            panic!("First argument must be of type Path");
-        };
-
-        let context_path = syn::parse_str("Context").unwrap();
-        if type_path.path != context_path {
-            panic!("First argument must be of type Context");
-        }
-
-        schema::Verb {
-            runtime: None,
-            pos: None,
-            comments: vec![],
-            export: false,
-            name: self.func.sig.ident.to_string(),
-            request: None,
-            response: None,
-            metadata: vec![],
-        }
-    }
-
-    pub fn get_request_type(&self) -> Box<syn::Type> {
-        let syn::FnArg::Typed(arg) = self.func.sig.inputs.last().unwrap() else {
-            panic!("Function must have two arguments");
-        };
-
-        arg.ty.clone()
-    }
-}
-
-/// Extract functions that are annotated with #[ftl::verb] and extract the AST node.
-pub fn extract_ast_verbs(module: &Ident, ast: syn::File) -> Vec<VerbToken> {
-    let ftl_verb_path = syn::parse_str("ftl::verb").unwrap();
-
-    let mut verbs = vec![];
-    for item in ast.items {
-        let item = match item {
-            syn::Item::Fn(func) => func,
-            _ => continue,
-        };
-
-        if !has_meta_path(&item.attrs, &ftl_verb_path) {
-            continue;
-        }
-
-        verbs.push(VerbToken {
-            module: module.clone(),
-            func: item,
-        });
-    }
-    verbs
-}
-
-// Look for #[path_str] e.g. #[ftl::verb], and extract the function signature.
-fn has_meta_path(attrs: &[syn::Attribute], expected_path: &Path) -> bool {
-    attrs.iter().any(|attr| attr.meta.path() == expected_path)
-}
-
-// TODO: make this a bit less overly specific. eg require_arg_type(&func, 0, "Context")
-fn fn_has_context_as_first_arg(func: &syn::ItemFn) -> bool {
-    let Some(arg) = func.sig.inputs.first() else {
-        println!("Function must have at least one argument");
-        return false;
-    };
-
-    let syn::FnArg::Typed(pat) = arg else {
-        println!("First argument must not be a self argument");
-        return false;
-    };
-
-    let syn::Type::Reference(path) = &*pat.ty else {
-        println!(
-            "First argument must be of type Reference instead of {:?}",
-            pat.ty
-        );
-        return false;
-    };
-    let syn::Type::Path(type_path) = &*path.elem else {
-        println!(
-            "First argument must be of type Path instead of {:?}",
-            path.elem
-        );
-        return false;
-    };
-
-    let context_path = syn::parse_str("Context").unwrap();
-    if type_path.path != context_path {
-        println!(
-            "First argument must be of type Context instead of {:?}",
-            type_path.path
-        );
-        return false;
-    }
-
-    true
 }
 
 #[cfg(test)]
